Add useDisconnect hook
